requirements.txt
Flask==3.0.0
Pillow==10.2.0
moviepy==1.0.3
numpy==1.24.3
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Requests==2.31.0
